esp_ml_training/requirements.txt
# Core ML and Data Handling
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
tensorflow>=2.8.0 # For TF/Keras model
tensorflowjs>=3.0.0 # For direct Keras to TF.js conversion using the Python library

# Data Ingestion (GCP)
google-cloud-firestore>=2.5.0
# google-cloud-bigquery>=2.30.0 # If BigQuery becomes a source

# Hyperparameter Tuning (Optional, examples)
# optuna>=2.10.0
# keras-tuner>=1.1.0

# Gradient Boosting (Alternative models)
xgboost>=1.5.0
lightgbm>=3.3.0

# Utility for saving/loading Python objects (like scalers)
joblib>=1.1.0

# Notebooks
jupyterlab>=3.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
